recipe_management/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-csp==4.0
django-filter==25.1
django-rename-app==0.1.7
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
mysql-connector-python==9.3.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
